+++ tests/test_ProtocolFileEncryptMultiThreaded.py
@@ -4,12 +4,16 @@ sys.path.append(".")
 
 # our package imports.
 #from smartinspectpython.siauto import *
+from smartinspectpython.siauto import SIAuto, SILevel, SISession, SIConfigurationTimer, SIColors
 
 print("Test Script Starting.\n")
 
-try:
+NUM_ENTRIES:int = 6000
 
-    from smartinspectpython.siauto import SIAuto, SILevel, SISession, SIConfigurationTimer, SIColors
+import threading
+
+# Function to simulate a task
+def print_numbers():
 
     # load SmartInspect settings from a configuration settings file.
     siConfigPath: str = "./tests/testdata/smartinspect_encrypted.cfg"
@@ -17,84 +21,33 @@ try:
 
     # start monitoring the configuration file for changes, and reload it when it changes.
     # this will check the file for changes every 60 seconds.
-    siConfig:SIConfigurationTimer = SIConfigurationTimer(SIAuto.Si, siConfigPath)
+    #siConfig:SIConfigurationTimer = SIConfigurationTimer(SIAuto.Si, siConfigPath)
 
     # get smartinspect logger reference; create a new session for this module name.
     _logsi:SISession = SIAuto.Si.GetSession(__name__)
     if (_logsi == None):
         _logsi = SIAuto.Si.AddSession(__name__, True)
-    _logsi.LogSeparator(SILevel.Error)
-    _logsi.LogVerbose("__init__.py HAS SpotifyPlus: initialization")
+    _logsi.LogVerbose("siThreadTest - See \"%s\" Watch Counter for results" % threading.current_thread().name)
     _logsi.LogAppDomain(SILevel.Verbose)
     _logsi.LogSystem(SILevel.Verbose)
 
-except Exception as ex:
-
-    print(str(ex))
-    #_LOGGER.warning("HAS SpotifyPlus could not init SmartInspect debugging! %s", str(ex))
-
-NUM_ENTRIES:int = 80000
-
-import threading
-import time
-
-# Function to simulate a task
-def print_numbers1():
     for i in range(1, NUM_ENTRIES):
-        _logsi.LogValue(SILevel.Debug, "print_numbers 1", i)
+        #_logsi.LogValue(SILevel.Debug, "print_numbers - %s" % threading.current_thread().name , i)
+        _logsi.IncCounter(SILevel.Debug, threading.current_thread().name)
 
-# Function to simulate another task
-def print_numbers2():
-    for i in range(2, NUM_ENTRIES):
-        _logsi.LogValue(SILevel.Debug, "print_numbers 2", i)
-
-# Function to simulate another task
-def print_numbers3():
-    for i in range(1, NUM_ENTRIES):
-        _logsi.LogValue(SILevel.Debug, "print_numbers 3", i)
-
-# Function to simulate another task
-def print_numbers4():
-    for i in range(1, NUM_ENTRIES):
-        _logsi.LogValue(SILevel.Debug, "print_numbers 4", i)
-
-# Function to simulate another task
-def print_numbers5():
-    for i in range(1, NUM_ENTRIES):
-        _logsi.LogValue(SILevel.Debug, "print_numbers 5", i)
-
-# Function to simulate another task
-def print_numbers6():
-    for i in range(1, NUM_ENTRIES):
-        _logsi.LogValue(SILevel.Debug, "print_numbers 6", i)
-
-# Function to simulate another task
-def print_numbers7():
-    for i in range(1, NUM_ENTRIES):
-        _logsi.LogValue(SILevel.Debug, "print_numbers 7", i)
-
-# Function to simulate another task
-def print_numbers8():
-    for i in range(1, NUM_ENTRIES):
-        _logsi.LogValue(SILevel.Debug, "print_numbers 8", i)
-
-# Function to simulate another task
-def print_numbers9():
-    for i in range(1, NUM_ENTRIES):
-        _logsi.LogValue(SILevel.Debug, "print_numbers 9", i)
 
 # Create threads
-thread1 = threading.Thread(target=print_numbers1, name="thread1")
-thread2 = threading.Thread(target=print_numbers2, name="thread2")
-thread3 = threading.Thread(target=print_numbers3, name="thread3")
-thread4 = threading.Thread(target=print_numbers4, name="thread4")
-thread5 = threading.Thread(target=print_numbers5, name="thread5")
-thread6 = threading.Thread(target=print_numbers6, name="thread6")
-thread7 = threading.Thread(target=print_numbers7, name="thread7")
-thread8 = threading.Thread(target=print_numbers8, name="thread8")
-thread9 = threading.Thread(target=print_numbers9, name="thread9")
+thread1 = threading.Thread(target=print_numbers, name="siTestThread1")
+thread2 = threading.Thread(target=print_numbers, name="siTestThread2")
+thread3 = threading.Thread(target=print_numbers, name="siTestThread3")
+thread4 = threading.Thread(target=print_numbers, name="siTestThread4")
+thread5 = threading.Thread(target=print_numbers, name="siTestThread5")
+thread6 = threading.Thread(target=print_numbers, name="siTestThread6")
+thread7 = threading.Thread(target=print_numbers, name="siTestThread7")
+thread8 = threading.Thread(target=print_numbers, name="siTestThread8")
+thread9 = threading.Thread(target=print_numbers, name="siTestThread9")
 
-# Start the threads
+# # Start the threads
 thread1.start()
 thread2.start()
 thread3.start()
@@ -105,7 +58,7 @@ thread7.start()
 thread8.start()
 thread9.start()
 
-# Wait for both threads to complete
+# # Wait for both threads to complete
 thread1.join()
 thread2.join()
 thread3.join()