+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 3.0.37 ] - 2025/03/27
+
+  * Fixed more bugs related to multi-threaded apps that log to encrypted log files.  Also added configuration file change monitoring automatically when the `SmartInspect.LoadConfiguration` method is called.
+
 ###### [ 3.0.36 ] - 2025/03/26
 
   * Fixed a bug with multi-threaded apps that log to encrypted log files; multiple log files were being created that contained an ending "a" suffix in the base file name.  The problem only seems to manifest itself when encryption options are specified.